+++ internal/service/repository/server.go
@@ -1,7 +1,6 @@
 package repository
 
 import (
-	"gitlab.com/gitlab-org/gitaly/internal/helper"
 	"gitlab.com/gitlab-org/gitaly/internal/rubyserver"
 
 	pb "gitlab.com/gitlab-org/gitaly-proto/go"
@@ -14,8 +13,4 @@ type server struct {
 // NewServer creates a new instance of a gRPC repo server
 func NewServer(rs *rubyserver.Server) pb.RepositoryServiceServer {
 	return &server{rs}
-}
-
-func (s *server) GetInfoAttributes(in *pb.GetInfoAttributesRequest, stream pb.RepositoryService_GetInfoAttributesServer) error {
-	return helper.Unimplemented
 }